+++ src/renderer/helpers/poolHelper.ts
@@ -17,16 +17,7 @@ import { PoolDetail } from '../types/generated/midgard'
 import { PoolTableRowData, PoolTableRowsData, PricePool } from '../views/pools/Pools.types'
 import { getPoolTableRowData } from '../views/pools/Pools.utils'
 import { convertBaseAmountDecimal, to1e8BaseAmount, isRuneAsset } from './assetHelper'
-import {
-  isBchChain,
-  isBnbChain,
-  isBtcChain,
-  isCosmosChain,
-  isDogeChain,
-  isEthChain,
-  isLtcChain,
-  isTerraChain
-} from './chainHelper'
+import { isBchChain, isBnbChain, isBtcChain, isCosmosChain, isDogeChain, isEthChain, isLtcChain } from './chainHelper'
 import { eqChain, eqString } from './fp/eq'
 import { ordBaseAmount } from './fp/ord'
 import { sequenceTOption, sequenceTOptionFromArray } from './fpHelpers'
@@ -200,7 +191,6 @@ export const disableAllActions = ({
     haltBtcChain,
     haltBchChain,
     haltDogeChain,
-    haltTerraChain,
     haltCosmosChain
   }
 }: {
@@ -229,9 +219,6 @@ export const disableAllActions = ({
   // Check `haltDogeChain` (provided by `mimir` endpoint) to disable all actions for DOGE pools
   if (isDogeChain(chain) && haltDogeChain) return true
 
-  // Check `haltTerraChain` (provided by `mimir` endpoint) to disable all actions for TERRA pools
-  if (isTerraChain(chain) && haltTerraChain) return true
-
   // Check `haltCosmosChain` (provided by `mimir` endpoint) to disable all actions for TERRA pools
   if (isCosmosChain(chain) && haltCosmosChain) return true
 
@@ -258,7 +245,6 @@ export const disableTradingActions = ({
     haltLtcTrading,
     haltBnbTrading,
     haltDogeTrading,
-    haltTerraTrading,
     haltCosmosTrading
   }
 }: {
@@ -276,7 +262,6 @@ export const disableTradingActions = ({
   if (isLtcChain(chain) && haltLtcTrading) return true
   if (isBnbChain(chain) && haltBnbTrading) return true
   if (isDogeChain(chain) && haltDogeTrading) return true
-  if (isTerraChain(chain) && haltTerraTrading) return true
   if (isCosmosChain(chain) && haltCosmosTrading) return true
 
   // 3. Check `chain` is included in `haltedChains` (provided by `inbound_addresses` endpoint)
@@ -294,17 +279,7 @@ export const disableTradingActions = ({
 export const disablePoolActions = ({
   chain,
   haltedChains,
-  mimirHalt: {
-    pauseLp,
-    pauseLpBnb,
-    pauseLpBch,
-    pauseLpBtc,
-    pauseLpEth,
-    pauseLpLtc,
-    pauseLpDoge,
-    pauseLpTerra,
-    pauseLpCosmos
-  }
+  mimirHalt: { pauseLp, pauseLpBnb, pauseLpBch, pauseLpBtc, pauseLpEth, pauseLpLtc, pauseLpDoge, pauseLpCosmos }
 }: {
   chain: Chain
   haltedChains: Chain[]
@@ -318,7 +293,6 @@ export const disablePoolActions = ({
   if (isEthChain(chain) && pauseLpEth) return true
   if (isLtcChain(chain) && pauseLpLtc) return true
   if (isDogeChain(chain) && pauseLpDoge) return true
-  if (isTerraChain(chain) && pauseLpTerra) return true
   if (isCosmosChain(chain) && pauseLpCosmos) return true
 
   // Check `chain` is included in `haltedChains` (provided by `inbound_addresses` endpoint)