+++ src/renderer/i18n/de/settings.ts
@@ -15,7 +15,9 @@ const settings: SettingMessages = {
   'setting.notconnected': 'Offline',
   'setting.add.device': 'Gerät hinzufügen',
   'setting.wallet.index': 'Index',
-  'setting.wallet.index.info': 'Trage die Index Nummer der Ledger Addresse ein, die Du verwenden möchtest'
+  'setting.wallet.index.info': 'Trage die Index Nummer der Ledger Addresse ein, die Du verwenden möchtest',
+  'setting.wallet.hdpath.legacy.info': 'Veralteter Derivation Pfad {path}',
+  'setting.wallet.hdpath.ledgerlive.info': 'Ledger Live Derivation Pfad {path}'
 }
 
 export default settings