+++ src/renderer/i18n/en/settings.ts
@@ -15,7 +15,9 @@ const settings: SettingMessages = {
   'setting.notconnected': 'Not connected',
   'setting.add.device': 'ADD DEVICE',
   'setting.wallet.index': 'Index',
-  'setting.wallet.index.info': 'Enter the index number of the Ledger address you want to use'
+  'setting.wallet.index.info': 'Enter the index number of the Ledger address you want to use',
+  'setting.wallet.hdpath.legacy.info': 'Legacy derivation path {path}',
+  'setting.wallet.hdpath.ledgerlive.info': 'Ledger Live derivation path {path}'
 }
 
 export default settings